+++ CHANGELOG.md
@@ -15,8 +15,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - OTLP data points re-use Resource and InstrumentationLibrary (thus are smaller). (#182)
-- Counters `sidecar.samples.produced` `sidecar.points.skipped` `sidecar.points.dropped`
-  and `sidecar.samples.processed` now account for histograms. (#187)
+
+### Removed
+
+- Removed counters `sidecar.samples.produced` & `sidecar.samples.processed`. (#187)
 
 ## [0.21.1](https://github.com/lightstep/opentelemetry-prometheus-sidecar/releases/tag/v0.21.1) - 2021-04-06